Refactor bounds rebuild into BoundsCacheSystem
