+++ viewer3d/viewer3dcontroller.cpp
@@ -58,6 +58,7 @@
 #include "render/opaque_fixture_pass.h"
 #include "render/opaque_object_pass.h"
 #include "render/opaque_truss_pass.h"
+#include "culling/bounds_cache_system.h"
 #include "culling/visibilitysystem.h"
 #include "labels/label_render_system.h"
 #include "picking/selectionsystem.h"
@@ -287,9 +288,6 @@ static std::array<float, 3> MakeDeterministicColor(std::string_view key) {
   return HsvToRgb(hue, sat, val);
 }
 
-static std::string ResolveCacheKey(const std::string &pathRef) {
-  return NormalizePath(pathRef);
-}
 
 static SymbolBounds ComputeSymbolBounds(const CommandBuffer &buffer) {
   SymbolBounds bounds{};
@@ -759,348 +757,14 @@ void Viewer3DController::UpdateResourcesIfDirty() {
     m_modelBounds.clear();
   }
 
-  RebuildBoundsIfDirty(hiddenLayers, trusses, objects, fixtures);
-}
-
-void Viewer3DController::RebuildBoundsIfDirty(
-    const std::unordered_set<std::string> &hiddenLayers,
-    const std::unordered_map<std::string, Truss> &trusses,
-    const std::unordered_map<std::string, SceneObject> &objects,
-    const std::unordered_map<std::string, Fixture> &fixtures) {
-  if (hiddenLayers != m_boundsHiddenLayers) {
-    Logger::Instance().Log("visibility dirty reason: hidden layers changed vs bounds cache");
-    m_visibilityChangedDirty = true;
-  }
-
-  if (!(m_sceneChangedDirty || m_assetsChangedDirty || m_visibilityChangedDirty) &&
-      m_cachedVersion == m_sceneVersion)
-    return;
-
-  if (m_visibilityChangedDirty) {
-    const auto eraseHidden = [&](auto &boundsMap, const auto &items) {
-      for (auto it = boundsMap.begin(); it != boundsMap.end();) {
-        auto itemIt = items.find(it->first);
-        if (itemIt == items.end() ||
-            !IsLayerVisibleCached(hiddenLayers, itemIt->second.layer)) {
-          it = boundsMap.erase(it);
-        } else {
-          ++it;
-        }
-      }
-    };
-    eraseHidden(m_fixtureBounds, fixtures);
-    eraseHidden(m_trussBounds, trusses);
-    eraseHidden(m_objectBounds, objects);
-    m_boundsHiddenLayers = hiddenLayers;
-    std::lock_guard<std::mutex> lock(m_sortedListsMutex);
-    m_sortedListsDirty = true;
-  }
-  m_cachedVersion = m_sceneVersion;
-
-  auto transformBounds = [](const Viewer3DController::BoundingBox &local,
-                            const Matrix &m) {
-    Viewer3DController::BoundingBox world;
-    world.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    world.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-    const auto &mn = local.min;
-    const auto &mx = local.max;
-    const std::array<std::array<float, 3>, 8> corners = {
-        std::array<float, 3>{mn[0], mn[1], mn[2]},
-        {mx[0], mn[1], mn[2]},
-        {mn[0], mx[1], mn[2]},
-        {mx[0], mx[1], mn[2]},
-        {mn[0], mn[1], mx[2]},
-        {mx[0], mn[1], mx[2]},
-        {mn[0], mx[1], mx[2]},
-        {mx[0], mx[1], mx[2]}};
-    for (const auto &c : corners) {
-      auto p = TransformPoint(m, c);
-      world.min[0] = std::min(world.min[0], p[0]);
-      world.min[1] = std::min(world.min[1], p[1]);
-      world.min[2] = std::min(world.min[2], p[2]);
-      world.max[0] = std::max(world.max[0], p[0]);
-      world.max[1] = std::max(world.max[1], p[1]);
-      world.max[2] = std::max(world.max[2], p[2]);
-    }
-    return world;
-  };
-  // Precompute bounding boxes for hover detection
-  m_fixtureBounds.clear();
-  for (const auto &[uuid, f] : fixtures) {
-    if (!IsLayerVisibleCached(hiddenLayers, f.layer))
-      continue;
-    Viewer3DController::BoundingBox bb;
-    Matrix fix = f.transform;
-    fix.o[0] *= RENDER_SCALE;
-    fix.o[1] *= RENDER_SCALE;
-    fix.o[2] *= RENDER_SCALE;
-
-    bool found = false;
-    bb.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    bb.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-
-    std::string gdtfPath;
-    auto gdtfPathIt = m_resourceSyncState.resolvedGdtfSpecs.find(ResolveCacheKey(f.gdtfSpec));
-    if (gdtfPathIt != m_resourceSyncState.resolvedGdtfSpecs.end() && gdtfPathIt->second.attempted)
-      gdtfPath = gdtfPathIt->second.resolvedPath;
-    auto itg = m_resourceSyncState.loadedGdtf.find(gdtfPath);
-    if (itg != m_resourceSyncState.loadedGdtf.end()) {
-      auto bit = m_modelBounds.find(gdtfPath);
-      if (bit == m_modelBounds.end()) {
-        BoundingBox local;
-        local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-        local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-        bool localFound = false;
-        for (const auto &obj : itg->second) {
-          for (size_t vi = 0; vi + 2 < obj.mesh.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {obj.mesh.vertices[vi] * RENDER_SCALE,
-                                      obj.mesh.vertices[vi + 1] * RENDER_SCALE,
-                                      obj.mesh.vertices[vi + 2] * RENDER_SCALE};
-            p = TransformPoint(obj.transform, p);
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-        }
-        if (localFound)
-          bit = m_modelBounds.emplace(gdtfPath, local).first;
-      }
-      if (bit != m_modelBounds.end()) {
-        bb = transformBounds(bit->second, fix);
-        found = true;
-      }
-    }
-
-    if (!found) {
-      float half = 0.1f;
-      std::array<std::array<float, 3>, 8> corners = {
-          std::array<float, 3>{-half, -half, -half},
-          {half, -half, -half},
-          {-half, half, -half},
-          {half, half, -half},
-          {-half, -half, half},
-          {half, -half, half},
-          {-half, half, half},
-          {half, half, half}};
-      for (const auto &c : corners) {
-        auto p = TransformPoint(fix, c);
-        bb.min[0] = std::min(bb.min[0], p[0]);
-        bb.min[1] = std::min(bb.min[1], p[1]);
-        bb.min[2] = std::min(bb.min[2], p[2]);
-        bb.max[0] = std::max(bb.max[0], p[0]);
-        bb.max[1] = std::max(bb.max[1], p[1]);
-        bb.max[2] = std::max(bb.max[2], p[2]);
-      }
-    }
-
-    m_fixtureBounds[uuid] = bb;
-  }
-
-  m_trussBounds.clear();
-  for (const auto &[uuid, t] : trusses) {
-    if (!IsLayerVisibleCached(hiddenLayers, t.layer))
-      continue;
-    BoundingBox bb;
-    Matrix tm = t.transform;
-    tm.o[0] *= RENDER_SCALE;
-    tm.o[1] *= RENDER_SCALE;
-    tm.o[2] *= RENDER_SCALE;
-
-    bool found = false;
-    bb.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    bb.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-
-    if (!t.symbolFile.empty()) {
-      std::string path;
-      auto pathIt = m_resourceSyncState.resolvedModelRefs.find(ResolveCacheKey(t.symbolFile));
-      if (pathIt != m_resourceSyncState.resolvedModelRefs.end() && pathIt->second.attempted)
-        path = pathIt->second.resolvedPath;
-      auto it = m_resourceSyncState.loadedMeshes.find(path);
-      if (it != m_resourceSyncState.loadedMeshes.end()) {
-        auto bit = m_modelBounds.find(path);
-        if (bit == m_modelBounds.end()) {
-          BoundingBox local;
-          local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-          local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-          bool localFound = false;
-          for (size_t vi = 0; vi + 2 < it->second.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {it->second.vertices[vi] * RENDER_SCALE,
-                                      it->second.vertices[vi + 1] * RENDER_SCALE,
-                                      it->second.vertices[vi + 2] * RENDER_SCALE};
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-          if (localFound)
-            bit = m_modelBounds.emplace(path, local).first;
-        }
-        if (bit != m_modelBounds.end()) {
-          bb = transformBounds(bit->second, tm);
-          found = true;
-        }
-      }
-    }
-
-    if (!found) {
-      float len = (t.lengthMm > 0 ? t.lengthMm * RENDER_SCALE : 0.3f);
-      float halfy = (t.widthMm > 0 ? t.widthMm * RENDER_SCALE * 0.5f : 0.15f);
-      float z1 = (t.heightMm > 0 ? t.heightMm * RENDER_SCALE : 0.3f);
-      std::array<std::array<float, 3>, 8> corners = {
-          std::array<float, 3>{0.0f, -halfy, 0.0f},
-          {len, -halfy, 0.0f},
-          {0.0f, halfy, 0.0f},
-          {len, halfy, 0.0f},
-          {0.0f, -halfy, z1},
-          {len, -halfy, z1},
-          {0.0f, halfy, z1},
-          {len, halfy, z1}};
-      for (const auto &c : corners) {
-        auto p = TransformPoint(tm, c);
-        bb.min[0] = std::min(bb.min[0], p[0]);
-        bb.min[1] = std::min(bb.min[1], p[1]);
-        bb.min[2] = std::min(bb.min[2], p[2]);
-        bb.max[0] = std::max(bb.max[0], p[0]);
-        bb.max[1] = std::max(bb.max[1], p[1]);
-        bb.max[2] = std::max(bb.max[2], p[2]);
-      }
-    }
-
-    m_trussBounds[uuid] = bb;
-  }
-
-  m_objectBounds.clear();
-  for (const auto &[uuid, obj] : objects) {
-    if (!IsLayerVisibleCached(hiddenLayers, obj.layer))
-      continue;
-    BoundingBox bb;
-    Matrix tm = obj.transform;
-
-    bool found = false;
-    bb.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    bb.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-
-    if (!obj.geometries.empty()) {
-      for (const auto &geo : obj.geometries) {
-        std::string path;
-        auto modelIt = m_resourceSyncState.resolvedModelRefs.find(ResolveCacheKey(geo.modelFile));
-        if (modelIt != m_resourceSyncState.resolvedModelRefs.end() && modelIt->second.attempted)
-          path = modelIt->second.resolvedPath;
-        auto it = m_resourceSyncState.loadedMeshes.find(path);
-        if (it == m_resourceSyncState.loadedMeshes.end())
-          continue;
-
-        auto bit = m_modelBounds.find(path);
-        if (bit == m_modelBounds.end()) {
-          BoundingBox local;
-          local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-          local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-          bool localFound = false;
-          for (size_t vi = 0; vi + 2 < it->second.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {it->second.vertices[vi] * RENDER_SCALE,
-                                      it->second.vertices[vi + 1] * RENDER_SCALE,
-                                      it->second.vertices[vi + 2] * RENDER_SCALE};
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-          if (localFound)
-            bit = m_modelBounds.emplace(path, local).first;
-        }
-        if (bit == m_modelBounds.end())
-          continue;
-
-        Matrix geoTm = MatrixUtils::Multiply(tm, geo.localTransform);
-        geoTm.o[0] *= RENDER_SCALE;
-        geoTm.o[1] *= RENDER_SCALE;
-        geoTm.o[2] *= RENDER_SCALE;
-        BoundingBox geoWorld = transformBounds(bit->second, geoTm);
-        bb.min[0] = std::min(bb.min[0], geoWorld.min[0]);
-        bb.min[1] = std::min(bb.min[1], geoWorld.min[1]);
-        bb.min[2] = std::min(bb.min[2], geoWorld.min[2]);
-        bb.max[0] = std::max(bb.max[0], geoWorld.max[0]);
-        bb.max[1] = std::max(bb.max[1], geoWorld.max[1]);
-        bb.max[2] = std::max(bb.max[2], geoWorld.max[2]);
-        found = true;
-      }
-    } else if (!obj.modelFile.empty()) {
-      std::string path;
-      auto modelIt = m_resourceSyncState.resolvedModelRefs.find(ResolveCacheKey(obj.modelFile));
-      if (modelIt != m_resourceSyncState.resolvedModelRefs.end() && modelIt->second.attempted)
-        path = modelIt->second.resolvedPath;
-      auto it = m_resourceSyncState.loadedMeshes.find(path);
-      if (it != m_resourceSyncState.loadedMeshes.end()) {
-        auto bit = m_modelBounds.find(path);
-        if (bit == m_modelBounds.end()) {
-          BoundingBox local;
-          local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-          local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-          bool localFound = false;
-          for (size_t vi = 0; vi + 2 < it->second.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {it->second.vertices[vi] * RENDER_SCALE,
-                                      it->second.vertices[vi + 1] * RENDER_SCALE,
-                                      it->second.vertices[vi + 2] * RENDER_SCALE};
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-          if (localFound)
-            bit = m_modelBounds.emplace(path, local).first;
-        }
-        if (bit != m_modelBounds.end()) {
-          Matrix scaledTm = tm;
-          scaledTm.o[0] *= RENDER_SCALE;
-          scaledTm.o[1] *= RENDER_SCALE;
-          scaledTm.o[2] *= RENDER_SCALE;
-          bb = transformBounds(bit->second, scaledTm);
-          found = true;
-        }
-      }
-    }
-
-    if (!found) {
-      float half = 0.15f;
-      std::array<std::array<float, 3>, 8> corners = {
-          std::array<float, 3>{-half, -half, -half},
-          {half, -half, -half},
-          {-half, half, -half},
-          {half, half, -half},
-          {-half, -half, half},
-          {half, -half, half},
-          {-half, half, half},
-          {half, half, half}};
-      for (const auto &c : corners) {
-        auto p = TransformPoint(tm, c);
-        bb.min[0] = std::min(bb.min[0], p[0]);
-        bb.min[1] = std::min(bb.min[1], p[1]);
-        bb.min[2] = std::min(bb.min[2], p[2]);
-        bb.max[0] = std::max(bb.max[0], p[0]);
-        bb.max[1] = std::max(bb.max[1], p[1]);
-        bb.max[2] = std::max(bb.max[2], p[2]);
-      }
-    }
-
-    m_objectBounds[uuid] = bb;
-  }
-
-  m_sceneChangedDirty = false;
-  m_assetsChangedDirty = false;
-  m_visibilityChangedDirty = false;
+  BoundsCacheSystem::Context boundsContext{
+      m_resourceSyncState, m_modelBounds, m_fixtureBounds,
+      m_trussBounds,      m_objectBounds, m_boundsHiddenLayers,
+      m_sceneVersion,     m_cachedVersion, m_sceneChangedDirty,
+      m_assetsChangedDirty, m_visibilityChangedDirty,
+      m_sortedListsMutex, m_sortedListsDirty};
+  BoundsCacheSystem::RebuildIfDirty(boundsContext, hiddenLayers, trusses,
+                                    objects, fixtures);
   m_lastHiddenLayers = hiddenLayers;
 }
